filter: Fix sample vs. tag propagation precision for fractional_* blocks
The fractional resampler and fractional interpolator blocks' output
samples would lose sync with tag propagation after a while due to
floating point precision problems:

- The blocks did their job with mixed float & double arithmetic
to generate output samples, but the runtime used double arithmetic
to propagate tags.

- Floating point reciprocals are quite exact in the least significant
bits, so specifying relative_rate as 1.0/resample_rate introduces
a mismatch in rates.

+++ gr-filter/lib/fractional_interpolator_ff_impl.h
@@ -33,8 +33,8 @@ namespace gr {
       : public fractional_interpolator_ff
     {
     private:
-      float d_mu;
-      float d_mu_inc;
+      double d_mu;
+      double d_mu_inc;
       mmse_fir_interpolator_ff *d_interp;
 
     public:
